backend/requirements.txt
aiofiles==23.2.1
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
audioread==3.0.1
bcrypt==4.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
ecdsa==0.19.1
fastapi==0.104.1
greenlet==3.2.4
groq==0.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.2
jiter==0.10.0
joblib==1.5.2
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.1
numba==0.61.2
numpy==2.2.6
openai==1.107.2
packaging==25.0
passlib==1.7.4
platformdirs==4.4.0
pluggy==1.6.0
pooch==1.8.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==5.0.1
requests==2.31.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
setuptools==58.0.4
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==1.0.0
SQLAlchemy==2.0.23
starlette==0.27.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
websockets==12.0
wheel==0.45.1
